+++ cypress/integration/lecturerCourseManagement.spec.js
@@ -10,6 +10,10 @@
  */
 
 describe("Course creation, edition and deletion", () => {
+    const random = Math.floor(Math.random() * 500);
+    const courseName = "test-course-cypress" + random;
+    const updatedCourseName = courseName + "-update";
+
     it("Login as lecturer", () => {
         cy.visit("/");
         cy.get("input[id='email']").type("lecturer");
@@ -24,7 +28,7 @@ describe("Course creation, edition and deletion", () => {
         cy.get("div[id='menu_courses']").children().eq(1).get("span").contains("All Courses").should("be.visible");
         cy.get("div[id='menu_courses']").children().eq(1).get("a").contains("All Courses").click();
         cy.get("div[id='menu_courses']").trigger("mouseleave");
-        cy.url().should("contain", "lecturer");
+        cy.url().should("contain", "course-management");
     }); 
 
     it("Show new course page", () => {
@@ -65,7 +69,7 @@ describe("Course creation, edition and deletion", () => {
     });
 
     it("Can edit courseName", () => {
-        cy.get('input[id="courseName"]').type("test-courseName-cypress");
+        cy.get('input[id="courseName"]').type(courseName);
     });
 
     it("Can edit courseLanguage", () => {
@@ -93,20 +97,20 @@ describe("Course creation, edition and deletion", () => {
 
     it("Create Course works", () => {
         cy.get('button[id="createCourse"]').click();
-        cy.url().should("contain", "lecturer");
+        cy.url().should("contain", "course-management");
     });
 
     it("Course was created", () => {
         cy.wait(3000);
         cy.get("button[title='Refresh']").click();
-        cy.get("div").contains("test-courseName-cypress");
+        cy.get("div").contains(courseName);
     });
 
     // edit course
     it("Show course edit page", () => {
-        cy.get("div").contains("test-courseName-cypress").parent().parent().find("button[id='editCourse']").click();
+        cy.get("div").contains(courseName).parent().parent().find("button[id='editCourse']").click();
 
-        cy.get('input[id="courseName"]').should("have.value", "test-courseName-cypress");
+        cy.get('input[id="courseName"]').should("have.value", courseName);
     });
 
     it("Can not save unchanged course", () => {
@@ -114,7 +118,7 @@ describe("Course creation, edition and deletion", () => {
     });
 
     it("Can edit courseName", () => {
-        cy.get('input[id="courseName"]').clear().type("test-CourseName-cypress");
+        cy.get('input[id="courseName"]').clear().type(updatedCourseName);
     });
 
     it("Can save course", () => {
@@ -125,12 +129,12 @@ describe("Course creation, edition and deletion", () => {
     it("Edit worked", () => {
         cy.wait(3000);
         cy.get("button[title='Refresh']").click();
-        cy.get("div").contains("test-CourseName-cypress").parent().parent().find("button[id='editCourse']").should("exist");
+        cy.get("div").contains(updatedCourseName).parent().parent().find("button[id='editCourse']").should("exist");
     });
 
     //delete course
     it("Show edit page", () => {
-        cy.get("div").contains("test-CourseName-cypress").parent().parent().find("button[id='editCourse']").click();
+        cy.get("div").contains(updatedCourseName).parent().parent().find("button[id='editCourse']").click();
     });
 
     it("Delete course", () => {
@@ -148,7 +152,7 @@ describe("Course creation, edition and deletion", () => {
     });
 
     it("Assert course deletion", () => {
-        cy.url().should("contain", "/lecturer");
-        cy.get("div").contains("test-CourseName-cypress").should("not.exist");
+        cy.url().should("contain", "/course-management");
+        cy.get("div").contains(updatedCourseName).should("not.exist");
     });
 });