+++ cypress/integration/studentCourseView.spec.js
@@ -1,4 +1,7 @@
 describe("Student course view", () => {
+    const random = Math.floor(Math.random() * 500);
+    const courseName = "test-course-cypress" + random;
+
     it("Create Course as lecturer", () => {
         cy.visit("/");
         cy.get("input[id='email']").type("lecturer");
@@ -8,16 +11,16 @@ describe("Student course view", () => {
         cy.get("div[id='menu_courses']").parents().eq(0).trigger("mouseover");
         cy.get("div[id='menu_courses']").children().eq(0).get("a").contains("All Courses").click();
         cy.get("div[id='menu_courses']").trigger("mouseleave");
-        cy.url().should("contain", "lecturer");
+        cy.url().should("contain", "course-management");
         // create course
         cy.get('button[id="addCourse"]').click({ force: true });
         cy.get("input[type='radio']").eq(0).click();
-        cy.get('input[id="courseName"]').type("test-courseName-cypress");
+        cy.get('input[id="courseName"]').type(courseName);
         cy.get("select").select("German");
         cy.get('textarea[id="courseDescription"]').type("test-courseDescription-cypress");
         cy.get('input[id="maxParticipants"]').clear().type("1");
         cy.get('button[id="createCourse"]').click();
-        cy.url().should("contain", "lecturer");
+        cy.url().should("contain", "course-management");
         cy.wait(1000);
     });
 
@@ -36,13 +39,13 @@ describe("Student course view", () => {
         cy.get("div[id='menu_courses']").children().eq(1).get("span").contains("All Courses").should("be.visible");
         cy.get("div[id='menu_courses']").children().eq(1).get("a").contains("All Courses").click();
         cy.get("div[id='menu_courses']").trigger("mouseleave");
-        cy.url().should("contain", "student");
+        cy.url().should("contain", "courses");
     }); 
 
     it("Course exists", () => {
         cy.wait(2000);
         cy.get("button[title='Refresh']").click();
-        cy.get("div").contains("test-courseName-cypress");
+        cy.get("div").contains(courseName);
         cy.reload();
     });
 
@@ -55,8 +58,8 @@ describe("Student course view", () => {
         cy.get("div[id='menu_courses']").parents().eq(0).trigger("mouseover");
         cy.get("div[id='menu_courses']").children().eq(0).get("a").contains("All Courses").click();
         cy.get("div[id='menu_courses']").trigger("mouseleave");
-        cy.url().should("contain", "lecturer");
-        cy.get("div").contains("test-courseName-cypress").parent().parent().find("button[id='editCourse']").click();
+        cy.url().should("contain", "course-management");
+        cy.get("div").contains(courseName).parent().parent().find("button[id='editCourse']").click();
         cy.get("button[id='deleteCourse']").click();
         cy.get('button[id="deleteCourseModalDelete"]').click();
     });