+++ src/components/SearchBar.vue
@@ -1,6 +1,6 @@
 <template>
     <div class="w-full relative text-gray-600 flex">
-        <button id="refresh" class="items-center mb-3" title="Search" @click="$emit('refresh')">
+        <button id="refresh" class="items-center mb-3" title="Search" @click="refresh">
             <i class="fas fa-search ml-3 absolute text-gray-600 hover:text-gray-700" />
         </button>
         <input
@@ -10,7 +10,7 @@
             type="search"
             placeholder="Filter"
         />
-        <button id="refresh" class="-ml-8 my-1 focus:outline-none" title="Refresh" @click="$emit('refresh')">
+        <button id="refresh" class="-ml-8 my-1 focus:outline-none" title="Refresh" @click="refresh">
             <i class="inline fas fa-redo-alt text-gray-600 hover:text-gray-700" />
         </button>
     </div>
@@ -18,6 +18,7 @@
 
 <script lang="ts">
     import { useModelWrapper } from "@/use/ModelWrapper";
+    import { ref } from "vue";
 
     export default {
         name: "SeachBar",
@@ -29,8 +30,26 @@
         },
         emits: ["refresh", "update:message"],
         setup(props: any, { emit }: any) {
+            let recentlyRefreshed = ref(false);
+            let filterMessage = ref("");
+
+            function preventRefresh() {
+                recentlyRefreshed.value = true;
+                setTimeout(() => {
+                    recentlyRefreshed.value = false;
+                }, 100);
+            }
+
+            function refresh() {
+                if (!recentlyRefreshed.value) {
+                    emit("refresh");
+                    preventRefresh();
+                }
+            }
+
             return {
                 filterMessage: useModelWrapper(props, emit, "message"),
+                refresh,
             };
         },
     };