+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# [v.0.6.1](https://github.com/upb-uc4/ui-web/compare/v0.6.0...v0.6.1) (2020-XX-XX)
+
+## Refactor
+- randomize test data to allow for concurrent testing [#353](https://github.com/upb-uc4/ui-web/pull/353)
+
+## Bugfix
+- add a timeout of 100ms to prevent refreshing the page too fast (caused application to crash) [#358](https://github.com/upb-uc4/ui-web/pull/358)
+
 # [v.0.6.0](https://github.com/upb-uc4/ui-web/compare/v0.5.1...v0.6.0) (2020-08-17)
 ## Feature
 - add filtering by role for the admin account list [#302](https://github.com/upb-uc4/ui-web/pull/302)
@@ -17,7 +25,7 @@
 - move course filtering for lecturer from frontend to backend [#317](https://github.com/upb-uc4/ui-web/pull/317)
 - remove unused wrapper components for asynchronous data loading [#341](https://github.com/upb-uc4/ui-web/pull/341)
 - replace placeholder descriptions in edit/create course form
-  
+
 ## Bugfix
 - Fix a possible XSS vulnerability [#141](https://github.com/upb-uc4/ui-web/issues/141)
 - Fix cursor styling of menus [#301](https://github.com/upb-uc4/ui-web/issues/301)